+++ demo/src/main/java/cn/bingoogolapple/refreshlayout/demo/activity/SwipeRecyclerViewDemoActivity.java
@@ -90,7 +90,7 @@ public class SwipeRecyclerViewDemoActivity extends AppCompatActivity implements 
     }
 
     @Override
-    public void onBGARefreshLayoutBeginLoadingMore(BGARefreshLayout refreshLayout) {
+    public boolean onBGARefreshLayoutBeginLoadingMore(BGARefreshLayout refreshLayout) {
         new AsyncTask<Void, Void, Void>() {
 
             @Override
@@ -109,6 +109,7 @@ public class SwipeRecyclerViewDemoActivity extends AppCompatActivity implements 
                 mAdapter.addDatas(DataEngine.loadMoreData());
             }
         }.execute();
+        return true;
     }
 
     @Override